linux-libc-headers-raspberrypi: Drop, its unneeded and bad practise
If we have this recipe, it means the whole armv6 (or whichever tune) is chosen depends
on a machine specific recipe. This makes no sense as armX packages are meant to be
machine independent.

We've had this problem in other layers and it causes needed rebuilds of packages
when you switch machines since the sstate checksums change. These headers are just part
of the toolchain bootstrap process so "standard" kernel headers are fine. The kernel
header version does need to be later of equal to the kernel version but we're fine in
that regard since the core is on 3.8, the latest pi kernel is 3.6.

There is nothing special about these headers so lets remove them and use the standard
system provided recipe, avoding any rebuilds. I tested the various other recipes in
the layer and there doesn't seem to be any dependency on these headers.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>